mqtt/requirements.txt
pandas
numpy
tensorflow
matplotlib
paho-mqtt